requirements.txt
Django==2.2.14
envparse==0.2.0
psycopg2==2.8.6
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0

celery==4.4.2
requests==2.23.0
django-appconf>=1.0.4
mock==4.0.3